18/11
Metodos criados
Entidade e derivadas:
LevaDano()=0
Danar () = 0

+++ JogoTecnicasProgramacao/Vilgax.h
@@ -8,14 +8,14 @@ namespace Entidades
             public Inimigo
         {
         protected:
-            int poder_dano;
+            float poder_dano;
         public:
             Vilgax();
             Vilgax(const sf::Vector2f pos);
             ~Vilgax();
 
-            void setPoderDano(int d);
-            int getPoderDano();
+            void setPoderDano(float d);
+            float getPoderDano();
 
         };
     }
